Changed values to current motors
Values considering the motor were changed.
The #defines were changed and instead of UM_PER_TICK there is now NM_PER_TICK (nanometer).

This needs to be watched carefully, because there could be an overflow. With a WHEEL_DIAMETER of 160mm the NM_PER_TICK is already at a value of 1'600'000'000. For a uint32_t this is close to the limit of about 4 billion.